Create pull request template (#192)
I looked into creating multiple PR templates and you have to use query
parameters to specify the template when creating a pull request, which
feels a little too complicated for now.

I think this PR template won't be relevant for all PRs, mostly for
component ones, but we can always simply edit it when creating a PR if
it's not relevant. So using it as the default makes sense to me for now.
We can adjust as we use it.

Thoughts on the emojis? lol. I'm not super great at picking emojis to
use so any input is welcome there.