feat(Modal): support disabled actions (#561)
## 📝 Changes

- Passes through `isDisabled` prop to action buttons to support them
being disabled
- Adds some documentation to Storybook for this but felt it didn't need
a full story or specific unit tests for this change

Note that with Easy UI we're moving to a more constraint-based design
for components to promote consistency. That's why we're not blindly
allowing any button prop to pass through the actions, because it can
lead to unintended Modal styles. So for now we're listing specific props
that are allowed to pass through.

## ✅ Checklist

- [x] Code is complete and in accordance with our style guide
- [x] Stories (docs) in Storybook accompany any relevant component
changes
- [x] Ensure no accessibility violations are reported in Storybook
- [x] Specs and documentation are up-to-date
- [x] Cross-browser check is performed (Chrome, Safari, Firefox)
- [x] Changeset is added