fix: clean up theme css vars (#901)
## 📝 Changes

- Update theme-aware tokens to remove `--ezui-t-` prefix and just use
`--ezui-` instead
- Removed theme-aware shadow tokens. This isn't clear if or how these
will be used in the future, so removing for now until we have an actual
use case
- Update "base" theme to "light" theme to be more clear with the theme's
intent/language used elsewhere in the ecosystem

This allows outside consumer to reference theme-aware CSS vars with a
less cryptic naming convention (i.e. `var(--ezui-color-primary-800)`
instead of `var(--ezui-t-color-primary-800)`—since the extra `t` is
kinda confusing). This will be more thoroughly documented in the near
future. Because the theme-aware tokens are aliases there should be no
collisions with global tokens, but this is something to be mindful of in
any future theme config.

I searched our ecosystem with hound and didn't see any uses of existing
references to our theme aware variables in other projects, so this
should be safe to release.

## ✅ Checklist

- [x] Visuals are complete and match Figma
- [x] Code is complete and in accordance with our style guide
- [x] Design and theme tokens are audited for any relevant changes
- [x] Unit tests are written and passing
- [x] TSDoc is written or updated for any component API surface area
- [x] Stories in Storybook accompany any relevant component changes
- [x] Ensure no accessibility violations are reported in Storybook
- [x] Specs and documentation are up-to-date
- [x] Cross-browser check is performed (Chrome, Safari, Firefox)
- [x] Changeset is added