Fix SASS deprecations (#1337)
## 📝 Changes

SASS introduced a [breaking
change](https://sass-lang.com/documentation/breaking-changes/mixed-decls/)
with nested selectors to be more in line with native CSS. This was
causing a lot of noise in our scripts.

This PR updates our SASS to remove the problematic code. Most of the
time the declartions in SCSS could simply be moved to the top of the
block. With complicated nesting (i.e. `_responsive-props.scss`) we have
to utilize `& { }`.

We'll need to be mindful of this SCSS change in new work—it may warrant
a Stylelint rule if it becomes problematic.

## ✅ Checklist

- [x] Code is complete and in accordance with our style guide
- [x] Changeset is added