react aria compatible menu selection (#1507)
- example of react aria menu compatible selection
- other menu fixes

Link to Story:

https://menu-adjustment--63f50c7c86f6514d2e0ef4be.chromatic.com/?path=/story/components-menu--multiple-selection-with-select-all

View "Actions" panel to see how behavior reflects what is expected for
React Aria selection. When selecting a list of items the returned keys
are an array of items. When selecting "Select all" directly, the key
returned is "all". This matches what is outlined here:
https://react-spectrum.adobe.com/react-aria/selection.html#select-all

We'll also want to update the docs to note this.