Fix: Fixes the DataGrid kebab button on safari (#1676)
## 📝 Changes

There is currently an issue in prod where on safari the kebab menu on
data grids do not work.

This just swaps out the `UnstyledPressButton` for `UnstyledButton` in
the kebab for the datagrid's kebab menu. Not sure why that fixes it, but
Steve suggested it and it does in fact fix it.

## ✅ Checklist

Easy UI has certain UX standards that must be met. In general,
non-trivial changes should meet the following criteria:

- [x] Visuals match Design Specs in Figma
- [x] Stories accompany any component changes
- [x] Code is in accordance with our style guide
- [x] Design tokens are utilized
- [x] Unit tests accompany any component changes
- [x] TSDoc is written for any API surface area
- [x] Specs are up-to-date
- [x] Console is free from warnings
- [x] No accessibility violations are reported
- [x] Cross-browser check is performed (Chrome, Safari, Firefox)
- [x] Changeset is added

~Strikethrough~ any items that are not applicable to this pull request.